marbleyou/requirements.txt
python-dotenv==1.0.0
asgiref==3.7.2
attrs==22.2.0
beautifulsoup4==4.11.2
colorama==0.4.6
crispy-bootstrap5==0.7
Django==4.2.6
django-bootstrap-form==3.4
django-bootstrap5==22.2
django-crispy-forms==2.1
django-debug-toolbar==4.2.0
django-phonenumber-field==7.2.0
djangorestframework==3.14.0
phonenumberslite == 8.13.22
Pillow==9.3.0
gunicorn==20.1.0
psycopg2-binary==2.9.1
